requirements.txt
functions-framework==3.*
pandas
google-cloud-storage
pandas-gbq
google-cloud-bigquery